## What's being changed

- explicitly casting floats to `int` before passing `microtime()` to `gmdate()` (this is used in all syncs)
- prevent `null` being passed to `number_format`
- check for valid array type before passing to `array_values()`

We've also:
- Removed a custom Dotdigitalgroup\Email\Model\Sql\Expression class (this was originally added for 2.1 support)
- Replace DateIntervalFactory with DateInterval [PR](https://dev.azure.com/dotdigital/ec/_git/ec-core-magento2-extension/pullrequest/35170)
- Fixed an incorrect constant path (product_link_style)
- Fixed the abandoned cart basket dynamic content endpoint
- Removed the `at()` matcher from our unit tests, to be compatible with PHPUnit 10

In most of the files we've touched, we've also repaired all errors surfaced by PHPStan and Magento2 coding standards.

## Why it's being changed

Magento 2.4.4 (coming early March) ships with _only_ PHP 8.1 support. Most of the errors we fixed relate to these two changes:
- https://wiki.php.net/rfc/deprecate_null_to_scalar_internal_arg
- https://wiki.php.net/rfc/implicit-float-int-deprecate

## How to review / test this change

All the component parts of this PR have been tested in their own smaller PRs. We'll do a full regression test prior to releasing.

Related work items: #158124, #165290

+++ Model/Customer/Guest.php
@@ -100,7 +100,7 @@ class Guest implements SyncInterface
         if ($this->countGuests) {
             $this->helper->log(
                 '----------- Guest sync ----------- : ' .
-                gmdate('H:i:s', microtime(true) - $this->start) .
+                gmdate('H:i:s', (int) (microtime(true) - $this->start)) .
                 ', Total synced = ' . $this->countGuests
             );
         }
