+++ Model/Sync/Review.php
@@ -172,7 +172,7 @@ class Review implements SyncInterface
         }
 
         $message = '----------- Review sync ----------- : ' .
-            gmdate('H:i:s', microtime(true) - $this->start) .
+            gmdate('H:i:s', (int) (microtime(true) - $this->start)) .
             ', synced = ' . $this->countReviews;
 
         if ($this->countReviews) {
